+++ release_notes.md
@@ -15,6 +15,7 @@
 - Add built-in support for the [Transparent Proxy](https://help.sap.com/docs/connectivity/sap-btp-connectivity-cf/transparent-proxy-for-kubernetes) via a new `TransparentProxyDestination`.
   For more information, refer to the [documentation](/docs/java/features/connectivity/transparent-proxy).
 - Add _experimental_ support for cross-level destination consumption via a new `CrossLevelScope` setting under `DestinationServiceOptionsAugmenter`.
+- Add _experimental_ support for setting custom headers in requests to the destination service via the new `DestinationServiceOptionsAugmenter#customHeaders`.
 
 ### 📈 Improvements